refactor: use @huddly/camera-proto and @huddly/camera-switch-proto packages
