+++ package.json
@@ -81,8 +81,8 @@
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-react": "^7.24.0",
     "generate-changelog": "^1.8.0",
-    "grpc-tools": "^1.11.1",
     "grpc_tools_node_protoc_ts": "^5.3.0",
+    "grpc-tools": "^1.11.1",
     "mocha": "^7.2.0",
     "nock": "^12.0.3",
     "nyc": "~15.0.1",
@@ -99,7 +99,8 @@
   },
   "dependencies": {
     "@grpc/grpc-js": "^1.3.2",
-    "@huddly/huddlyproto": "0.0.3",
+    "@huddly/camera-proto": "^1.0.2",
+    "@huddly/camera-switch-proto": "^0.0.1",
     "cpio-stream": "^1.4.1",
     "google-protobuf": "^3.17.3",
     "jszip": "3.2.2",
