+++ src/interfaces/ICameraSwitchModels.ts
